Extract displaying five latest bids into separate component
